refactor: replace setTheme from Html to lib function
